+++ bin/tmctl.sh
@@ -589,7 +589,7 @@ cmd_server_edit() {
 }
 
 cmd_version() {
-    echo "TimeMachine Backup v3.6.12"
+    echo "TimeMachine Backup v3.7.0"
 }
 
 cmd_fix_permissions() {